Add staircase function to print pattern
Implement staircase function to print a staircase pattern based on input integer.

Staircase Example:  Here is a staircase of size n = 4:
      #
    ##
  ###
####

Its base and height are both equal to n. It is drawn using # symbols and spaces. The last line is not preceded by any spaces.

Write a program that prints a staircase of size n.

Function Description

Complete the staircase function with the following parameter(s):  int 'n' = type integer.

Print a staircase as described above. No value should be returned from the method.
Note: The last line is not preceded by spaces. All lines are right-aligned.